+++ rust/dbn/src/encode/csv.rs
@@ -37,7 +37,8 @@ where
         }
     }
 
-    fn encode_header<R: super::DbnEncodable>(&mut self) -> anyhow::Result<()> {
+    #[doc(hidden)]
+    pub fn encode_header<R: super::DbnEncodable>(&mut self) -> anyhow::Result<()> {
         R::serialize_header(&mut self.writer)?;
         // end of line
         self.writer.write_record(None::<&[u8]>)?;